fix: move deal with icon codes to a single file
