+++ frontend/rust-lib/flowy-folder2/src/entities/mod.rs
@@ -1,9 +1,11 @@
+pub mod icon;
 mod import;
 mod parser;
 pub mod trash;
 pub mod view;
 pub mod workspace;
 
+pub use icon::*;
 pub use import::*;
 pub use trash::*;
 pub use view::*;
