+++ frontend/rust-lib/flowy-folder2/tests/workspace/script.rs
@@ -1,5 +1,6 @@
 use collab_folder::core::ViewLayout;
 
+use flowy_folder2::entities::icon::{UpdateViewIconPayloadPB, ViewIconPB};
 use flowy_folder2::entities::*;
 use flowy_folder2::event_map::FolderEvent::*;
 use flowy_test::event_builder::EventBuilder;
